requirements.txt
about-time==4.2.1
absl-py==1.3.0
alive-progress==3.1.0
altgraph==0.17.3
asgiref==3.6.0
astor==0.8.1
auto-py-to-exe==2.34.0
beautifulsoup4==4.11.1
bleach==1.5.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
Cython==0.29.34
html5lib==0.9999999
lxml==4.9.2
telebot==0.0.5
PyTelegramBotApi==4.12.0
flask==2.0.0
psycopg==3.1.12
